requirements.txt
langchain>=0.1.20,<0.3.0
langchain-community>=0.0.38,<0.3.0
langchain-groq
chromadb==0.5.0
sentence-transformers>=2.7.0
pypdf
pymupdf
python-dotenv
fastapi
uvicorn[standard]
python-multipart
lark>=1.3.1
